requirements.txt
torch
torchvision
pyyaml
pillow
tqdm
pytest
black
flake8
mlflow
hyfra-core